Move crown layers used by backward_cloning only in decomon.models
Convert2Mode being used by both forward and backward cloning goes to
decomon.models.utils

merge_with_previous() function being used only there goes in
decomon.models.crown
